Resolves #53, by being able to choose wether or not the form is a registration form. Standard is that it is a registration form.
