+++ code/JS/formCreator.js
@@ -45,6 +45,7 @@ function HTXJS_updateForm(formid) {
         formid: formid,
         tableName: $('#tableName').val(),
         tableDescription: $('#tableDescription').val(),
+        registration: $('#registrationInput').is(":checked") ? 1 : 0,
         arrived: $('#arrivedInput').is(":checked") ? 1 : 0,
         crew: $('#crewInput').is(":checked") ? 1 : 0,
         pizza: $('#pizzaInput').is(":checked") ? 1 : 0,
